Aggiornamento a .netCore 3.1 e implementazione di .NET Identity

+++ MyCourse/Models/InputModels/Lessons/LessonDeleteInputModel.cs
@@ -6,6 +6,6 @@ namespace MyCourse.Models.InputModels.Lessons
     {
         [Required]
         public long id { get; set; }
-        public int CourseId { get; set; }
+        public long CourseId { get; set; }
     }
 }
